requirements.txt
streamlit
torch
torchvision
Pillow
transformers
ftfy
regex
tqdm
gtts
sentencepiece
streamlit-lottie